fix: display settings page, add dashboard header
