+++ app/(dashboard)/dashboard/layout.tsx
@@ -1,21 +1,25 @@
-import React from "react";
+"use client";
+
+import { useState } from "react";
 import Link from "next/link";
 import { Icons } from "@/components/ui/icons";
 import DashboardHeader from "@/components/dashboard/DashboardHeader";
 import DashboardActions from "@/components/dashboard/DashboardActions";
+import Settings from "@/app/(dashboard)/settings/page";
 
 const Layout = ({ children }: { children: React.ReactNode }) => {
+  const [showSettings, setShowSettings] = useState(false);
+
   return (
     <>
       <main className="px-6 py-4 w-full max-w-7xl mx-auto flex flex-col main-h-svh ">
         <div className="space-y-8 w-full h-full">
-          {/* Header/Navbar */}
           <DashboardHeader />
+          
+          {showSettings && <Settings />}
 
-          {/* Main Section - Table */}
           <div className="space-y-0 w-full ">
             <div className="w-full flex justify-between max-[800px]:flex-col-reverse">
-              {/* Tabs List/Component (Stock, Sales, Report) */}
               <div className="w-full flex">
                 <Link href={"/dashboard"}>
                   <div className="flex items-center justify-center gap-2 border border-b-white py-2 rounded-tr-lg rounded-tl-lg w-44 max-[800px]:w-full font-semibold px-9 shadow-inner">
@@ -37,14 +41,12 @@ const Layout = ({ children }: { children: React.ReactNode }) => {
                 </Link>
               </div>
 
-              {/* Add New Stock button and Search Component */}
               <DashboardActions />
             </div>
             {children}
           </div>
         </div>
 
-        {/* Footer */}
         <div className="flex flex-col gap-2 mt-4">
           <p className="text-center mt-4">
             &copy; {new Date().getFullYear()}, Powered by Timbu Business
